requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
kociemba==1.2.1
groq==0.13.0
python-dotenv==1.0.1
pydantic==2.10.3
